Refactor MainControllerTest
