+++ src/test/java/pro/sky/animal_shelter_telegram_bot/controller/ConstantsForControllerTests.java
@@ -1,11 +1,7 @@
 package pro.sky.animal_shelter_telegram_bot.controller;
 
 import pro.sky.animal_shelter_telegram_bot.model.PetOwner;
-import pro.sky.animal_shelter_telegram_bot.model.Report;
 import pro.sky.animal_shelter_telegram_bot.model.pets.Pet;
-
-import java.util.ArrayList;
-import java.util.Collection;
 
 public class ConstantsForControllerTests {
 
@@ -20,7 +16,6 @@ public class ConstantsForControllerTests {
 
     public final static Pet PET = new Pet();
     public final static PetOwner PET_OWNER = new PetOwner();
-    public final static Collection<Report> REPORTS = new ArrayList<>();
 
     public final static long ID = 1L;
     public final static String FIRST_NAME = "Ivan";
@@ -34,8 +29,6 @@ public class ConstantsForControllerTests {
     public final static String DATE_OF_REPORT = "12.12.2021";
     public final static String DIET = "Meat";
     public final static String RESULT = "Report is fine";
-
-    public final static String HELLO_MESSAGE_MAIN_CONTROLLER = "Welcome to our pet shelter!";
 
     public final static String HELLO_MESSAGE_PET_CONTROLLER = "You can do it by information of pet:\n" +
             "1. add pet information\n" +
